(#28)  Rename the slashes submodule to commands.  Hoist import from fixture to module in test_{knobs,jobads,submits}.
